% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch a Marxan app}
\usage{
launch_app(appname)
}
\arguments{
\item{appname}{An app name.}
}
\description{
Launch a Marxan app
}
\examples{
launch_app("import")
launch_app("manage")
launch_app("marxan")
launch_app("marzone")
launch_app("mxptest")
}

