% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMarxan.R
\name{runMarxan}
\alias{runMarxan}
\title{Launch a Marxan app}
\usage{
runMarxan(appname)
}
\arguments{
\item{appname}{An app name.}
}
\description{
Launch a Marxan app
}
\examples{
runMarxan("import")
runMarxan("manage")
runMarxan("marxan")
runMarxan("marzone")
runMarxan("mxptest")
}

